+++ package.json
@@ -14,6 +14,7 @@
     "bs58": "^5.0.0",
     "dotenv": "^16.4.1",
     "pino": "^8.18.0",
+    "pino-multi-stream": "^6.0.0",
     "pino-pretty": "^10.3.1",
     "pino-std-serializers": "^6.2.2"
   },
