+++ utils/logger.ts
@@ -15,3 +15,32 @@ export const logger = pino(
   },
   transport,
 );
+
+
+// const transport = pino.transport({
+//   targets: [
+//     {
+//       level: 'trace',
+//       target: 'pino/file',
+//       options: {
+//         destination: 'bot.log',
+//         // customPrettifiers: {
+//         // },
+//         ignore: 'pid,hostname,level',
+//       },
+//     },
+//     {
+//       level: 'trace',
+//       target: 'pino-pretty',
+//       options: {},
+//     },
+//   ],
+// });
+
+// export const logger = pino(transport);
+
+
+// const customPrettifier = (log: any) => {
+//   const { level, time, pid, hostname, ...rest } = JSON.parse(log);
+//   return `${new Date(time).toISOString()} ${JSON.stringify(rest)}`;
+// };
